requirements.txt
streamlit
chromadb
langchain
sentence-transformers
pypdf
ollama
